Respect android:turnScreenOn on non-default displays
PowerManager#wakeUp always wakes up only the default display (and
all displays that are in the default display group). However,
virtual device displays are in their own non-default group, so
the wakeUp calls from WM need to include the relevant displayId.

Note that if that displayId is doesn't belong to a virtual device
(i.e. it's also in the default display group, this is a no-op).

Fix: 365175880
Test: CTS
Flag: EXEMPT trivial bugfix
Change-Id: Ib078db24f6bced42f78416a18b124cc5b4a4e374