Reload RBC config when colorMode changes.
RBC uses two matrixes depending on color mode: linear or gamma.
When color mode was changing for color correction, we weren't
updating the RBC matrix and this resulted in extra-dark visuals.

This CL updates RBC coefficients when colorMode changes.

Also had to add some changes to set the RBC matrix early
when colorMode does change to reduce the lag. Without this
lag was resulting in some jarring flicker when color
correct was toggled.

Bug: 336354594
Test: atest ColorDisplayServiceTest, DisplayTransformManagerTest
Flag: EXEMPT bugfix
Change-Id: I5e29325c46f6ee716c11acf006b817bbc4da88d4