[PiP2] Update DisplayLayout on remote rotation too
We need to make sure display layout is properly rotated
when we receive a remote display rotation, not just
display change as a part of requestStartTransition().

We also need to make sure we do not consider this case as
fixed-rotation, by checking both start and endRotation.
If the PiP change has a well defined endRotation it means
the orientation flipped during collection and we won't be
doing async rotation, so handle this case as normal (same-orientation)
enter PiP animation case.

Bug: 378506208
Flag: com.android.wm.shell.enable_pip2
Test: enter PiP2 in fixed rotation with multi-activity app in ges-nav
Change-Id: I116396d27d4d5e6520fbc1d9e17f7a115b32d7a6