Listen to bold text setting change in onConfigurationChanged
The text in the magnification setting panel will be bolded when the
setting changes and the panel is opened, but the text won't change if
the panel is closed and reopened. This is becasue TextView will update
itself when there is a configuration change, but when the panel is
closed and TextViews are not drawn, text won't be updated.

Therefore, we need to listen to the onConfigurationChanged signal for it
in the panel.

Bug: 380229091
Test: Manually. Add video in the bug.
Flag: EXEMPT bugfix
Change-Id: I41a5cec85b67cfcbc16d4a62fe1dd097cad8028b