Do not re-use dead shared isolated processes.
If we are trying to bind a service into an existing isolated process shortly after a previous shared process has been killed, we may still find the existing ProcessRecord, because that is only cleaned up after we receive a death notification. When then trying to bring up the service in that process, we notice it is dead, and schedule a restart, which introduces an additional delay into the binding.

Instead, make sure we don't use a ProcessRecord if it has already been killed.

Bug: 380037707
Change-Id: Iea591fad9113f522dae1bcd4ccbd1adc6fbb1076
Test: attest CtsExternalServiceTestCases
Flag: EXEMPT bugfix