LocationFudger: cache must fetch default when missing
The new density-based coarse locations are not used if the cache doesn't have a default coarsening value.
The cache fetches the default when created, however, before this commit, if this initial fetch fails, it never retries.
This changes the LocationFudger behavior so that, if the feature flag is enabled, each time the cache is needed, if the default is not present, it is re-fetched.

Tests:
- atest FrameworksMockingServicesTests:LocationFudgerTest
- atest FrameworksMockingServicesTests:LocationFudgerCacheTest

Test: manual atest on Pixel 7 pro (see above)

Flag: android.location.flags.density_based_coarse_locations

Bug: 381816602

Change-Id: I77ad32802cf6101edb7a269f5c006add2808a751