[2/2][PiP2] Operate on TF token for deferConfig
Pass TaskFragment token separately in a new PipChange
parcelable if the PiP candidate activity is in an embedded TaskFragment.

Using this token for enter PiP WCT, will make sure that in the AE case,
activities in other TFs aren't marked as config-at-end with a paused
dispatch.

Bug: 380025892
Flag: com.android.wm.shell.enable_pip2
Test: atest SecondaryActivityEnterPipTest
Change-Id: I280492aa8d3446683e9d910b1ac0077f2d5ce54a