Add back gesture navigation to Glanceable Hub.
This changelist adds back gesture handling on Glanceable Hub, snapping
back to the lockscreen when invoked.

This change also removes a communal tutorial interactor test that
is predicated on setting the tutorial state before entering the
Glanceable Hub. This is incompatible with a state flow for isShowing
as the default initial state will immediately set the tutorial to
shown. The tutorial code is unused and deprecated so the test has
been removed.

Flag: com.android.systemui.glanceable_hub_back_action
Test: atest CommunalBackActionInteractorTest
Fixes: 382771533
Change-Id: I3e50b52aa12cd27f20a7260385c3d4b534f4e65b