[flexiglass] Large clock not clipping away
Fixes layout and clipping issue that was causing the step animation that
pushes the clock digits to the left to make the clock invisible.

Two parts:
a. Let the clock composable be as wide as its parent composable, don't
2x shrink it anymore, the step animation will take care shifting the
digits off to the left when notifications are shown
b. Disable clipping in the AndroidView-wrapped View hierarchy of the
large clock

Two additional bugs were discovered while fixing this one:
1. On tablet in portrait mode, the clock is cut off. See b/384077475
2. The step animation works well when animating from left to right but
   "jumps" when animating from the center to the left. See b/384082024

Fix: 370045350
Test: manually verified on tablet that the large clock is visible with
and without a notification and correctly "step animates" depending on
the notifications being there or not
Test: manually verified the same on an unfolded foldable
Flag: com.android.systemui.scene_container

Change-Id: Ic616fb73fbfe727183e40f4dc899c614c974045a