[Dual Shade] Set the shade invocation zone to always be split 50/50.
This is due to a recently changed decision that followed from moving
notifications to the left. Instead of a dynamic invocation zone based
on the position of status bar notifications icons, we'll have a simple
50/50 split of the status bar to invoke the notifications shade (on the
left) or quick settings shade (on the right).

This will be mirrored in RTL layouts respectively.

For the time being, we will keep the code in place to make this dynamic
again in the future.

Fix: 384898612
Test: Manually tested by opening and closing both the notifications
 shade and quick settings shade on a folded and unfolded device.
Flag: com.android.systemui.dual_shade
Change-Id: Ib74cab24e6cfc64f1988d1002bfdfe7dd6a86d6b