Add Navigation Component
Add a navigationComponent field to AppJankStats. This field is intended
to be populated with a navigation destination, screen or pane name. This
iteration arose after a discussion with the Compose team where it was
identified that the concept of a Widget_Id does not really apply to
composable functions.

Bug: 385173526
Test: atest CoreAppJankTestCases
flag: android.app.jank.detailed_app_jank_metrics_api
Change-Id: I32ad7a398fcf73352eed7aa48a8210707d08a198