Add flag for porting window-resize animation to shell transit
This flag is for porting the WindowChangeAnimationSpec logic
from legacy transitions to shell transitions. This is
supposed to be the default animation (which covers for
*->fullscreen).

This seems to have been an oversight -- however, since
this technically can add animation to where there wasn't
one, put it behind a flag to be safe. Want to use it
for bubble-anything, so porting it now.

Bug: 384976265
Test: N/A - just a flag
Flag: com.android.window.flags.port_window_size_animation
Change-Id: I8d68d8d687d73866a1514b6e61961531fa90a8ad