Fix the format of argument name comments
This CL updates the format of argument name comments.

The following command is used to find comments to be fixed:

  rg '/\*\s+\w+\s+\*/' \
    -g "**/desktopmode/**" \
    -g "!src/com/android/wm/shell/desktopmode/DesktopModeEventLogger.kt" \
    -g "!tests/unittest/src/com/android/wm/shell/desktopmode/DesktopModeEventLoggerTest.kt"

The comments are then replaced with the `/* name= */ value` style, or
Kotlin's named arguments.

Some comments in DesktopModeEventLogger.kt and
DesktopModeEventLoggerTest.kt are unchanged. This is because the proto
binding generator gives arguments less meaningful names like `arg0`.
Using `/* name= */ value` style comments with inconsistent names
resulted in warnings in the IDE.

Bug: None
Test: None
Flag: DOCS_ONLY
Change-Id: I626ee944b13f99ccd4b23536c21fd1aac3ece08f