Fix IME flicker when dismissing stack
When the bubbles stack is collapsed or expanded and the IME is visible,
we request to hide it and start the expand or collapse animation at
the same time. This causes a flicker.

This change first hides the IME and waits for it to be hidden before
starting the expand or collapse animations.

Demo: http://recall/-/bJtug1HhvXkkeA4MQvIaiP/cxDr6MipIW3dC5By64LchE

Flag: EXEMPT bug fix
Bug: 379474417
Test: atest BubbleStackViewTest
Change-Id: Ic14b97b40a6482abf93b22c48be99639bf177824