Separate task-view management out of taskview impl [1/N]
Start of separating taskview control functionality out
of TaskView impl.

There are a number of calls in TaskView where the taskview
is asking a "higher-level" to perform a general management
operation on it (eg. "please resize me" or "please close me"
in a transition). In practice, though, its the "higher-level"
that is asking the taskview to do this, so we end up with a
sorta inside-out logic where the outer controller asks the
taskview to do something which then just goes back to the
controller to ask it to do the thing.

This means that some of the overall taskview control is
implemented within the TaskView, itself. As a result, we
can't substitute the taskview control implementation for
different use-cases (eg. bubbles vs auto) which we need
in order to support transitions tailored to each interaction.

To remedy this, the plan is to "lift" the taskview control
functionality into an interface that can have different
implementations. The refactor is mechanical, but will be
separated into a few CLs to make it easier to see that it
is mechanical.

The mechanical operations in this CL are:
- verbatim move calls like startActivity from TaskViewTaskController
  to TaskViewTransitions
- Have TaskView call the moved calls on TaskViewTransitions instead.
- BubbleController makes its own instance of TaskViewTransitions
  when shared repository is enabled (behind flag).
- Removed some unused functions

Bug: 384976265
Test: refactor only, so existing tests
Flag: EXEMPT mechanical refactor
Change-Id: I31951b20a3c6b618620e0eb110ba94d6bb2bbd18