[SB] Fix StatusBarSignalPolicyTest when ethernet flag is on.
When the status_bar_signal_policy_refactor_ethernet flag is on,
`StatusBarIconController#setIconVisibility` will be called for ethernet.
This updates StatusBarSignalPolicyTest to only make assertions on
specific slot calls, so that tests will pass even if we start calling
`#setIconVisibility(slotEthernet)`.

See b/291321279#comment25 for the test failure when trying to enable the
ethernet flag.

Bug: 291321279
Test: atest StatusBarSignalPolicyTest
Flag: TEST_ONLY
Change-Id: I1b343898b0f2e6ca6924431d296986f2649a7874