Update policy cache when clearing policy
Permission grant policy getter returns cached value while setter relies
on the underlying value in DevicePolicyData to check if there is a
change. As a result if they diverge, it is possible that the setter call
is ignored because the value is the same as in DevicePolicyData, while
cached value remains different.

Bug: 376650409
Test: atest CtsDevicePolicyManagerTestCases:com.android.cts.devicepolicy.MixedProfileOwnerTest
Flag: EXEMPT bugfix
Change-Id: I00d457da1a7337f91a1e34e3d8cf8b63ea548516